move expectedResult from TxResultsTests to TxTests
