+++ src/transactions/TxResultsTests.cpp
@@ -34,88 +34,6 @@ using namespace stellar::txtest;
 namespace
 {
 
-struct ExpectedOpResult
-{
-    OperationResultCode code;
-    OperationType type;
-    CreateAccountResultCode createAccountCode;
-    PaymentResultCode paymentCode;
-    AccountMergeResultCode accountMergeCode;
-    SetOptionsResultCode setOptionsResultCode;
-
-    ExpectedOpResult(OperationResultCode code) : code{code}
-    {
-    }
-    ExpectedOpResult(CreateAccountResultCode createAccountCode)
-        : code{opINNER}
-        , type{CREATE_ACCOUNT}
-        , createAccountCode{createAccountCode}
-    {
-    }
-    ExpectedOpResult(PaymentResultCode paymentCode)
-        : code{opINNER}, type{PAYMENT}, paymentCode{paymentCode}
-    {
-    }
-    ExpectedOpResult(AccountMergeResultCode accountMergeCode)
-        : code{opINNER}, type{ACCOUNT_MERGE}, accountMergeCode{accountMergeCode}
-    {
-    }
-    ExpectedOpResult(SetOptionsResultCode setOptionsResultCode)
-        : code{opINNER}
-        , type{SET_OPTIONS}
-        , setOptionsResultCode{setOptionsResultCode}
-    {
-    }
-};
-
-TransactionResult
-expectedResult(int64_t fee, size_t opsCount, TransactionResultCode code,
-               std::vector<ExpectedOpResult> ops = {})
-{
-    auto result = TransactionResult{};
-    result.feeCharged = fee;
-    result.result.code(code);
-    if (code != txSUCCESS && code != txFAILED)
-    {
-        return result;
-    }
-    if (ops.empty())
-    {
-        std::fill_n(std::back_inserter(ops), opsCount, PAYMENT_SUCCESS);
-    }
-
-    result.result.results().resize(static_cast<uint32_t>(ops.size()));
-    for (size_t i = 0; i < ops.size(); i++)
-    {
-        auto& r = result.result.results()[i];
-        auto& o = ops[i];
-        r.code(o.code);
-        if (o.code == opINNER)
-        {
-            r.tr().type(o.type);
-            switch (o.type)
-            {
-            case CREATE_ACCOUNT:
-                r.tr().createAccountResult().code(o.createAccountCode);
-                break;
-            case PAYMENT:
-                r.tr().paymentResult().code(o.paymentCode);
-                break;
-            case ACCOUNT_MERGE:
-                r.tr().accountMergeResult().code(o.accountMergeCode);
-                break;
-            case SET_OPTIONS:
-                r.tr().setOptionsResult().code(o.setOptionsResultCode);
-                break;
-            default:
-                break;
-            }
-        }
-    }
-
-    return result;
-}
-
 enum class Signed
 {
     NOT_SIGNED,
@@ -168,12 +86,6 @@ variations(int count, std::vector<T> values)
     return result;
 }
 
-struct ValidationResult
-{
-    int64_t fee;
-    TransactionResultCode code;
-};
-
 TEST_CASE("txresults", "[tx][txresults]")
 {
     auto const& cfg = getTestConfig();
@@ -182,44 +94,12 @@ TEST_CASE("txresults", "[tx][txresults]")
     auto app = createTestApplication(clock, cfg);
     app->start();
 
-    auto validate = [&](TransactionFramePtr const& tx,
-                        ValidationResult validationResult,
-                        TransactionResult const& applyResult = {}) {
-        auto shouldValidateOk = validationResult.code == txSUCCESS;
-        REQUIRE(tx->checkValid(*app, 0) == shouldValidateOk);
-        REQUIRE(tx->getResult().result.code() == validationResult.code);
-        REQUIRE(tx->getResult().feeCharged == validationResult.fee);
-
-        // do not try to apply if checkValid returned false
-        if (!shouldValidateOk)
-        {
-            REQUIRE(applyResult == TransactionResult{});
-            return;
-        }
-
-        switch (applyResult.result.code())
-        {
-        case txINTERNAL_ERROR:
-        case txBAD_AUTH_EXTRA:
-        case txBAD_SEQ:
-            return;
-        default:
-            break;
-        }
-
-        auto shouldApplyOk = applyResult.result.code() == txSUCCESS;
-        auto applyOk = applyCheck(tx, *app);
-        REQUIRE(tx->getResult() == applyResult);
-        REQUIRE(applyOk == shouldApplyOk);
-    };
-
     auto& lm = app->getLedgerManager();
     auto& clh = lm.getCurrentLedgerHeader();
     clh.scpValue.closeTime = 10;
     const int64_t baseReserve = clh.baseReserve;
     const int64_t baseFee = clh.baseFee;
     const int64_t startAmount = baseReserve * 100;
-    const int64_t paymentAmount = baseReserve * 10;
 
     auto amount = [&](PaymentValidity t) {
         switch (t)
@@ -382,7 +262,7 @@ TEST_CASE("txresults", "[tx][txresults]")
             {
                 auto tx = a.tx({});
                 for_all_versions(*app, [&] {
-                    validate(tx, {0, txMISSING_OPERATION});
+                    validateTxResults(tx, *app, {0, txMISSING_OPERATION});
                 });
             }
 
@@ -392,7 +272,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().tx.timeBounds.activate().minTime =
                     clh.scpValue.closeTime + 1;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txTOO_EARLY});
+                    validateTxResults(tx, *app, {baseFee, txTOO_EARLY});
                 });
             }
 
@@ -402,7 +282,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().tx.timeBounds.activate().maxTime =
                     clh.scpValue.closeTime - 1;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txTOO_LATE});
+                    validateTxResults(tx, *app, {baseFee, txTOO_LATE});
                 });
             }
 
@@ -411,7 +291,8 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = a.tx({payment(root, 1)});
                 tx->getEnvelope().tx.fee--;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee - 1, txINSUFFICIENT_FEE});
+                    validateTxResults(tx, *app,
+                                      {baseFee - 1, txINSUFFICIENT_FEE});
                 });
             }
 
@@ -419,7 +300,7 @@ TEST_CASE("txresults", "[tx][txresults]")
             {
                 auto tx = f.tx({payment(root, 1)});
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txNO_ACCOUNT});
+                    validateTxResults(tx, *app, {baseFee, txNO_ACCOUNT});
                 });
             }
 
@@ -428,7 +309,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = a.tx({payment(root, 1)});
                 tx->getEnvelope().tx.seqNum++;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txBAD_SEQ});
+                    validateTxResults(tx, *app, {baseFee, txBAD_SEQ});
                 });
             }
 
@@ -436,7 +317,8 @@ TEST_CASE("txresults", "[tx][txresults]")
             {
                 auto tx = g.tx({payment(root, 1)});
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txINSUFFICIENT_BALANCE});
+                    validateTxResults(tx, *app,
+                                      {baseFee, txINSUFFICIENT_BALANCE});
                 });
             }
         }
@@ -448,7 +330,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = a.tx({});
                 tx->getEnvelope().signatures.clear();
                 for_all_versions(*app, [&] {
-                    validate(tx, {0, txMISSING_OPERATION});
+                    validateTxResults(tx, *app, {0, txMISSING_OPERATION});
                 });
             }
 
@@ -459,7 +341,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().tx.timeBounds.activate().minTime =
                     clh.scpValue.closeTime + 1;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txTOO_EARLY});
+                    validateTxResults(tx, *app, {baseFee, txTOO_EARLY});
                 });
             }
 
@@ -470,7 +352,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().tx.timeBounds.activate().maxTime =
                     clh.scpValue.closeTime - 1;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txTOO_LATE});
+                    validateTxResults(tx, *app, {baseFee, txTOO_LATE});
                 });
             }
 
@@ -480,7 +362,8 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().signatures.clear();
                 tx->getEnvelope().tx.fee--;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee - 1, txINSUFFICIENT_FEE});
+                    validateTxResults(tx, *app,
+                                      {baseFee - 1, txINSUFFICIENT_FEE});
                 });
             }
 
@@ -489,7 +372,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = f.tx({payment(root, 1)});
                 tx->getEnvelope().signatures.clear();
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txNO_ACCOUNT});
+                    validateTxResults(tx, *app, {baseFee, txNO_ACCOUNT});
                 });
             }
 
@@ -499,7 +382,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().signatures.clear();
                 tx->getEnvelope().tx.seqNum++;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txBAD_SEQ});
+                    validateTxResults(tx, *app, {baseFee, txBAD_SEQ});
                 });
             }
 
@@ -508,13 +391,14 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = g.tx({payment(root, 1)});
                 tx->getEnvelope().signatures.clear();
                 for_versions_to(6, *app, [&] {
-                    validate(tx, {baseFee, txBAD_AUTH});
+                    validateTxResults(tx, *app, {baseFee, txBAD_AUTH});
                 });
                 for_versions({7}, *app, [&] {
-                    validate(tx, {baseFee, txINSUFFICIENT_BALANCE});
+                    validateTxResults(tx, *app,
+                                      {baseFee, txINSUFFICIENT_BALANCE});
                 });
                 for_versions_from(8, *app, [&] {
-                    validate(tx, {baseFee, txBAD_AUTH});
+                    validateTxResults(tx, *app, {baseFee, txBAD_AUTH});
                 });
             }
         }
@@ -526,7 +410,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = a.tx({});
                 tx->addSignature(a);
                 for_all_versions(*app, [&] {
-                    validate(tx, {0, txMISSING_OPERATION});
+                    validateTxResults(tx, *app, {0, txMISSING_OPERATION});
                 });
             }
 
@@ -537,7 +421,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().tx.timeBounds.activate().minTime =
                     clh.scpValue.closeTime + 1;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txTOO_EARLY});
+                    validateTxResults(tx, *app, {baseFee, txTOO_EARLY});
                 });
             }
 
@@ -548,7 +432,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->getEnvelope().tx.timeBounds.activate().maxTime =
                     clh.scpValue.closeTime - 1;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txTOO_LATE});
+                    validateTxResults(tx, *app, {baseFee, txTOO_LATE});
                 });
             }
 
@@ -558,7 +442,8 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->addSignature(a);
                 tx->getEnvelope().tx.fee--;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee - 1, txINSUFFICIENT_FEE});
+                    validateTxResults(tx, *app,
+                                      {baseFee - 1, txINSUFFICIENT_FEE});
                 });
             }
 
@@ -567,7 +452,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = f.tx({payment(root, 1)});
                 tx->addSignature(a);
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txNO_ACCOUNT});
+                    validateTxResults(tx, *app, {baseFee, txNO_ACCOUNT});
                 });
             }
 
@@ -577,7 +462,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 tx->addSignature(a);
                 tx->getEnvelope().tx.seqNum++;
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txBAD_SEQ});
+                    validateTxResults(tx, *app, {baseFee, txBAD_SEQ});
                 });
             }
 
@@ -586,7 +471,8 @@ TEST_CASE("txresults", "[tx][txresults]")
                 auto tx = g.tx({payment(root, 1)});
                 tx->addSignature(a);
                 for_all_versions(*app, [&] {
-                    validate(tx, {baseFee, txINSUFFICIENT_BALANCE});
+                    validateTxResults(tx, *app,
+                                      {baseFee, txINSUFFICIENT_BALANCE});
                 });
             }
         }
@@ -642,7 +528,7 @@ TEST_CASE("txresults", "[tx][txresults]")
                 signs, ops, app->getLedgerManager().getCurrentLedgerVersion());
             auto applyResult = makeApplyResult(
                 signs, ops, app->getLedgerManager().getCurrentLedgerVersion());
-            validate(tx, validationResult, applyResult);
+            validateTxResults(tx, *app, validationResult, applyResult);
         });
     };
 
@@ -684,7 +570,8 @@ TEST_CASE("txresults", "[tx][txresults]")
                 .accountMergeResult()
                 .sourceAccountBalance() = startAmount - 1200;
             for_all_versions(*app, [&] {
-                validate(tx, {baseFee * 2, txSUCCESS}, applyResult);
+                validateTxResults(tx, *app, {baseFee * 2, txSUCCESS},
+                                  applyResult);
             });
         }
 
@@ -694,8 +581,9 @@ TEST_CASE("txresults", "[tx][txresults]")
                 a.tx({payment(b, 1000), accountMerge(root), payment(c, 1000)});
 
             for_all_versions(*app, [&] {
-                validate(tx, {baseFee * 3, txSUCCESS},
-                         expectedResult(baseFee * 3, 3, txINTERNAL_ERROR));
+                validateTxResults(
+                    tx, *app, {baseFee * 3, txSUCCESS},
+                    expectedResult(baseFee * 3, 3, txINTERNAL_ERROR));
             });
         }
     }
@@ -707,9 +595,9 @@ TEST_CASE("txresults", "[tx][txresults]")
             auto tx = root.tx({createAccount(f, startAmount)});
 
             for_all_versions(*app, [&] {
-                validate(tx, {baseFee * 1, txSUCCESS},
-                         expectedResult(baseFee * 1, 1, txSUCCESS,
-                                        {CREATE_ACCOUNT_SUCCESS}));
+                validateTxResults(tx, *app, {baseFee * 1, txSUCCESS},
+                                  expectedResult(baseFee * 1, 1, txSUCCESS,
+                                                 {CREATE_ACCOUNT_SUCCESS}));
             });
         }
 
@@ -720,8 +608,8 @@ TEST_CASE("txresults", "[tx][txresults]")
             tx->addSignature(a);
 
             for_all_versions(*app, [&] {
-                validate(
-                    tx, {baseFee * 2, txSUCCESS},
+                validateTxResults(
+                    tx, *app, {baseFee * 2, txSUCCESS},
                     expectedResult(baseFee * 2, 2, txSUCCESS,
                                    {CREATE_ACCOUNT_SUCCESS, PAYMENT_SUCCESS}));
             });
@@ -742,8 +630,8 @@ TEST_CASE("txresults", "[tx][txresults]")
                 a.tx({payment(b, 1000), setOptions(nullptr, nullptr, nullptr,
                                                    &th, nullptr, nullptr)});
             for_all_versions(*app, [&] {
-                validate(
-                    tx, {baseFee * 2, txSUCCESS},
+                validateTxResults(
+                    tx, *app, {baseFee * 2, txSUCCESS},
                     expectedResult(baseFee * 2, 2, txSUCCESS,
                                    {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS}));
             });
@@ -757,22 +645,25 @@ TEST_CASE("txresults", "[tx][txresults]")
                  payment(c, 1000)});
 
             for_versions_to(6, *app, [&] {
-                validate(tx, {baseFee * 3, txSUCCESS},
-                         expectedResult(baseFee * 3, 3, txFAILED,
-                                        {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS,
-                                         opBAD_AUTH}));
+                validateTxResults(
+                    tx, *app, {baseFee * 3, txSUCCESS},
+                    expectedResult(
+                        baseFee * 3, 3, txFAILED,
+                        {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS, opBAD_AUTH}));
             });
             for_versions({7}, *app, [&] {
-                validate(tx, {baseFee * 3, txSUCCESS},
-                         expectedResult(baseFee * 3, 3, txSUCCESS,
-                                        {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS,
-                                         PAYMENT_SUCCESS}));
+                validateTxResults(
+                    tx, *app, {baseFee * 3, txSUCCESS},
+                    expectedResult(baseFee * 3, 3, txSUCCESS,
+                                   {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS,
+                                    PAYMENT_SUCCESS}));
             });
             for_versions_from(8, *app, [&] {
-                validate(tx, {baseFee * 3, txSUCCESS},
-                         expectedResult(baseFee * 3, 3, txFAILED,
-                                        {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS,
-                                         opBAD_AUTH}));
+                validateTxResults(
+                    tx, *app, {baseFee * 3, txSUCCESS},
+                    expectedResult(
+                        baseFee * 3, 3, txFAILED,
+                        {PAYMENT_SUCCESS, SET_OPTIONS_SUCCESS, opBAD_AUTH}));
             });
         }
 
@@ -782,15 +673,15 @@ TEST_CASE("txresults", "[tx][txresults]")
             auto tx = a.tx({payment(b, 1000)});
 
             for_versions_to(6, *app, [&] {
-                validate(tx, {baseFee * 1, txFAILED});
+                validateTxResults(tx, *app, {baseFee * 1, txFAILED});
             });
             for_versions({7}, *app, [&] {
-                validate(tx, {baseFee * 1, txSUCCESS},
-                         expectedResult(baseFee * 1, 1, txSUCCESS,
-                                        {PAYMENT_SUCCESS}));
+                validateTxResults(tx, *app, {baseFee * 1, txSUCCESS},
+                                  expectedResult(baseFee * 1, 1, txSUCCESS,
+                                                 {PAYMENT_SUCCESS}));
             });
             for_versions_from(8, *app, [&] {
-                validate(tx, {baseFee * 1, txFAILED});
+                validateTxResults(tx, *app, {baseFee * 1, txFAILED});
             });
         }
     }
